+++ mainProject/parent/imagerecognition/platerecognition/docker/buildDockerImage.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 # A sample Bash script, to build docker file
 echo Building docker image
-docker build -t messagehandler:latest .
+docker build -t platerecognition:latest .
