+++ mainProject/parent/imagerecognition/platerecognition/docker/pom.xml
@@ -3,7 +3,7 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>solent.ac.uk.devops.traffic</groupId>
-        <artifactId>messagehandlerexampleparent</artifactId>
+        <artifactId>platerecognition</artifactId>
         <version>0.1-SNAPSHOT</version>
     </parent>
     
@@ -61,7 +61,7 @@
         <!-- project one-jar dependencies -->
         <dependency>
             <groupId>solent.ac.uk.devops.traffic</groupId>
-            <artifactId>messagehandler-assembly</artifactId>
+            <artifactId>platerecognition-assembly</artifactId>
             <version>${project.version}</version>
             <scope>compile</scope>
             <classifier>onejar</classifier>
