+++ mainProject/parent/imagerecognition/platerecognition/service/pom.xml
@@ -3,7 +3,7 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>solent.ac.uk.devops.traffic</groupId>
-        <artifactId>messagehandlerexampleparent</artifactId>
+        <artifactId>platerecognition</artifactId>
         <version>0.1-SNAPSHOT</version>
     </parent>
     <artifactId>platerecognition-service</artifactId>
