+++ mainProject/parent/imagerecognition/platerecognition/service/src/main/java/org/solent/devops/traffic/platerecognition/Controller.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.solent.devops.traffic.messageexample;
+package org.solent.devops.traffic.platerecognition;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
