+++ mainProject/parent/imagerecognition/platerecognition/service/src/test/java/org/solent/devops/traffic/platerecognition/test/JMSJUnitTest.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.solent.devops.traffic.messageexample.test;
+package org.solent.devops.traffic.platerecognition.test;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -15,7 +15,7 @@ import org.junit.Test;
 import static org.junit.Assert.*;
 import org.junit.runner.RunWith;
 import org.solent.devops.message.jms.SimpleJmsSender;
-import org.solent.devops.traffic.messageexample.Controller;
+import org.solent.devops.traffic.platerecognition.Controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
